chore(Context selector): deprecate the Context Selector component
