+++ packages/react-core/src/components/index.ts
@@ -21,7 +21,6 @@ export * from './Chip';
 export * from './ChipGroup';
 export * from './ClipboardCopy';
 export * from './CodeBlock';
-export * from './ContextSelector';
 export * from './DataList';
 export * from './DatePicker';
 export * from './DescriptionList';
