+++ packages/react-core/src/deprecated/components/index.ts
@@ -1,1 +1,2 @@
 export * from './OptionsMenu';
+export * from './ContextSelector';
